% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_function_to_imputed_data.R
\name{apply_function_to_imputed_data}
\alias{apply_function_to_imputed_data}
\title{Apply function to dataframes in a mice object}
\usage{
apply_function_to_imputed_data(mice_data, fun, ...)
}
\arguments{
\item{mice_data}{a mids object generated by \code{mice::mice()}.}

\item{fun}{the function to apply to each dataframe. May only take one
positional argument of type data.frame.}

\item{...}{other arguments passed to fun()}
}
\value{
a mids object with transformed data.
}
\description{
Wrapper function to apply a function on each dataframe in an imputed dataset
created with \code{mice::mice()}.
}
\author{
J. Peter Marquardt
}
