% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_formula_tools.R
\name{deconstruct_formula}
\alias{deconstruct_formula}
\title{Deconstruct formula}
\usage{
deconstruct_formula(formula)
}
\arguments{
\item{formula}{formula object for use in statistical models}
}
\value{
a named list with fields
* outcome (character)
* predictors (vector of characters)
* censor_event (character) (optional) censor event, only for formulas including a Surv() object
}
\description{
Deconstruct a formula object into strings of its components
}
\examples{
deconstruct_formula(outcome ~ predictor1 + predictor2 + predictor3)
deconstruct_formula(Surv(outcome, censor_event) ~ predictor)

}
\author{
J. Peter Marquardt
}
