% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_names_types_factors.R
\name{.find_NA_coercions}
\alias{.find_NA_coercions}
\title{Locate NA values introduced during apply_data_dictionary()}
\usage{
.find_NA_coercions(data_raw, data, data_dictionary)
}
\arguments{
\item{data_raw}{data.frame that was provided as input to \code{apply_data_dictionary()}}

\item{data}{data.frame that is returned by \code{apply_data_dictionary()}}

\item{data_dictionary}{the data_dictionary used by \code{apply_data_dictionary()}
to turn "data_raw" into "data".}
}
\value{
returns a dataframe with the location of introduced NA's. If no NA's
were introduced an empty dataframe is returned.
}
\description{
Finds and locates NA values that were introduced by calling \code{apply_data_dictionary()}
on a dataframe using a data_dictionary.
}
\author{
Till D. Best, J. Peter Marquardt
}
\keyword{internal}
