% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_names_types_factors.R
\name{.parse_string_to_named_vector}
\alias{.parse_string_to_named_vector}
\title{Parse a string to create a named list}
\usage{
.parse_string_to_named_vector(str)
}
\arguments{
\item{str}{character with standardised pattern to be parsed}
}
\value{
named vector
}
\description{
Create a named list from a standardised string of the follwoing format:
\itemize{
\item key-value pairs are separated from other key-value-pairs by a comma (",")
\item key and value of the same pair are separated by an equal sign ("=")
\item quotations around individual keys and values are recommended for clarity, but do not affect functionality.
\item all values will be coerced to type character, with the exception of "NA", "TRUE" and "FALSE"
}
}
\examples{
basecamb:::.parse_string_to_named_vector("'key1' = 'val1', 'key2' = 'val2'")
basecamb:::.parse_string_to_named_vector("'key1' = 'val1', 'key2' = 'val2', 'default' = 'NA'")

}
\author{
J. Peter Marquardt
}
\keyword{internal}
