% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_mult_impute_observed_outcome.R
\name{.remove_missing_from_mids}
\alias{.remove_missing_from_mids}
\title{Remove missing cases from a mids object}
\usage{
.remove_missing_from_mids(mids, var)
}
\arguments{
\item{mids}{mids objects that is filtered.}

\item{var}{a string or vector of strings specifying the variable(s). All
cases (i.e. rows) for which there are missing values are removed.}
}
\value{
a mids object filtered for observed cases of var.
}
\description{
!!! Deprecated !!! This used to be a helper function for basecamb::fit_mult_impute_obs_outcome
that is no longer needed. !!! Deprecated !!!
}
\details{
Remove_missing_from_mids is used to filter a mids object for missing cases
in the original dataset in the variable var. This is useful for situations
where you want to use as many observations as possible for imputation but
only fit your model on a subset of these. Or, if you want to create one
large imputed datset from which multiple analyses with multiple outcomes
are derived.
}
\author{
Till D. Best
}
