% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_continuous_predictors.R
\name{.scale_variable}
\alias{.scale_variable}
\title{Scaling a variable}
\usage{
.scale_variable(data, variable, scaling_denominator)
}
\arguments{
\item{data}{data.frame}

\item{variable}{a char indicating the variable to be scaled}

\item{scaling_denominator}{a numeric indicating the scaling. The variable is
divided by the scaling_denominator.}
}
\value{
the input dataframe with the newly scaled 'variable'
}
\description{
A helper function to scale a variable in a dataframe.
Divides 'variable' by 'scaling_denominator'.
}
