% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_model_summary.R
\name{or_model_summary}
\alias{or_model_summary}
\title{Summarise a logistic/ordinal regression model on the odds ratio scal in a dataframe}
\usage{
or_model_summary(
  model,
  conf_int = 1.96,
  print_intercept = FALSE,
  round_est = 3,
  round_p = 4
)
}
\arguments{
\item{model}{a model object with estimates on the log-odds scale.}

\item{conf_int}{a numeric used to calculate the confidence intervals. The
default of 1.96 gives the 95\% confidence interval.}

\item{print_intercept}{a logical flag indicating whether intercepts shall
be removed. Intercept removing is done with a regular expression matching
"y>=". If you have a variable matching this pattern it will also be removed!}

\item{round_est}{the number of decimals provided for estimates (odds ratios)
and confidence intervals. If the input has decimals, it will be
rounded to the nearest integer.}

\item{round_p}{the number of decimals provided for p-values. If the input has
decimals, it will be rounded to the nearest integer.}
}
\value{
a dataframe with the adjusted odds ratio, confidence intervals and
  p-values.
}
\description{
This function summarises regression models that return data on the log-odds
  scale and returns a dataframe with estimates, and confidence intervals as
  odds ratios. P value are also provided.
  There's also the option to remove intercepts. This comes in handy when you
  fit proportional odds ordinal logistic regression models which usually
  result in many intercepts that are not really of interest.
  This function also works with models obtained from multiply imputed
  datasets, for example fitted with Hmisc::fit.mult.impute().
}
\details{
CAVE! The function does not check whether you're estimates are on the
  log-odds scale. It 'only' performs the transformation!
}
\examples{
\dontrun{
fit_mult_impute_obs_outcome(mids = imputed_data, formula = "y ~ x1 + x2", fitter = orm)
}

}
\author{
Till D. Best
}
