% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_continuous_predictors.R
\name{scale_continuous_predictors}
\alias{scale_continuous_predictors}
\title{Scale continuous predictors}
\usage{
scale_continuous_predictors(data, scaling_dictionary)
}
\arguments{
\item{data}{a data object with variables.}

\item{scaling_dictionary}{a data.frame with two columns that are called
"variable" and "scaling_denominator".}
}
\value{
The data with the newly scaled 'variables'.
}
\description{
This function linearly scales variables in data objects according to a data dictionary.
The data dictionary has at least two columns, "variable" and "scaling_denominator".
"Variable" is divided by "scaling_denominator".
}
\author{
Till D. Best
}
