% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicates.R
\name{setduplicates}
\alias{setduplicates}
\title{Identify duplicate values in a vector representing a set}
\usage{
setduplicates(vect)
}
\arguments{
\item{vect}{a vector of any type}
}
\value{
a vector of duplicate elements
}
\description{
Identify duplicate values in a vector representing a set
}
\examples{
setduplicates(c(1,2,2,3))

}
\seealso{
\link[generics]{setops}
}
\author{
J. Peter Marquardt
}
