% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_function_to_imputed_data.R
\name{apply_function_to_imputed_data}
\alias{apply_function_to_imputed_data}
\title{Apply function to dataframes in a mice object}
\usage{
apply_function_to_imputed_data(mice_data, fun)
}
\arguments{
\item{mice_data}{a mids object generated by mice()}

\item{fun}{the function to apply to each data.frame. May only take one positional argument of type data.frame}
}
\value{
a mids object with transformed data
}
\description{
Wrapper function to transform the imputed data.frames within a mids object generated by mice()
}
\author{
J. Peter Marquardt
}
