% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mult_impute_obs_outcome.R
\name{fit_mult_impute_obs_outcome}
\alias{fit_mult_impute_obs_outcome}
\title{Fit a proportional odds model on multiply imputed data}
\usage{
fit_mult_impute_obs_outcome(imp, outcome, predictors, fitter = orm)
}
\arguments{
\item{imp}{mids object containing multiple imputations of data generated with
mice::mice()}

\item{outcome}{the outcome column as a string}

\item{predictors}{a vector of strings containing the predictors/covariables}

\item{fitter}{the fitter used for fitting the model. For details see}
}
\value{
est the pooled estimate
}
\description{
This function fits an proportional odds logistic regression (a type of
  ordinal logistic regression) on a multiply imputed dataset that was
  generated with mice::mice(). Cases with a missing outcome before imputation
  are removed from the fitting process.
}
\author{
Till D. Best
}
