% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_mult_impute_observed_outcome.R
\name{fit_mult_impute_obs_outcome}
\alias{fit_mult_impute_obs_outcome}
\title{Fit a model on multiply imputed data using only non-missing outcomes}
\usage{
fit_mult_impute_obs_outcome(mids, formula, fitter, ...)
}
\arguments{
\item{mids}{a mids object, i.e. the imputed dataset}

\item{formula}{a formula object that describes the model to be fit}

\item{fitter}{the fitter (i.e. the type of model) used by
Hmisc::fit.mult.impute to fit the model.}

\item{...}{any other arguments to Hmisc::fit.mult.impute()}
}
\value{
mod a fit.mult.impute object
}
\description{
This function fits a regression model using Hmisc::fit.mult.impute() on a
  multiply imputed dataset generated with mice::mice(). Cases with a
  missing outcome are removed from the dataset before model fitting.
}
\examples{
\dontrun{
fit_mult_impute_obs_outcome(mids = imputed_data, formula = "y ~ x1 + x2", fitter = rms::orm)
}

}
\author{
Till D. Best
}
